OV-198: * move control component to studio components
