+++ frontend/src/bundles/studio/components/components.ts
@@ -1,3 +1,4 @@
+export { Control } from './control/control.js';
 export { PlayerControls } from './player-controls/player-controls.js';
 export { Timeline } from './timeline/timeline.js';
 export { VideoComponent } from './video/video.js';
