+++ frontend/src/bundles/studio/components/player-controls/components/components.ts
@@ -1,2 +1,1 @@
-export { Control } from './control/control.js';
 export { TimeDisplay } from './time-display/time-display.js';
