+++ frontend/src/bundles/studio/components/timeline/rows/scenes-row.tsx
@@ -1,12 +1,12 @@
 import { Badge, Image } from '~/bundles/common/components/components.js';
 import { useAppSelector, useMemo } from '~/bundles/common/hooks/hooks.js';
 import { IconName, IconSize } from '~/bundles/common/icons/icons.js';
+import { Control } from '~/bundles/studio/components/components.js';
 import { RowNames } from '~/bundles/studio/enums/enums.js';
 import { setItemsSpan } from '~/bundles/studio/helpers/set-items-span.js';
 import { useTimelineContext } from '~/bundles/studio/hooks/hooks.js';
 import styles from '~/framework/theme/styles/css-modules/timeline.module.css';
 
-import { Control } from '../../player-controls/components/components.js';
 import { Item, Row } from '../components.js';
 
 const ScenesRow: React.FC = () => {
